+++ lib/example/example_file.dart
@@ -1,29 +1,27 @@
 import 'package:anitube_crawler_api/anitube_crawler_api.dart';
 
 void main() async {
-  var localTimeout = 7000; // defining 7 seconds timeout
-  AniTubeApi api = AniTubeApi();
-  var homePage;
-  try {
-    // This call returns a HomePageInfo instance or throws an exception
-    // if somethings goes wrong,
-    homePage = await api.getHomePageData(timeout: localTimeout);
+  // var localTimeout = 7000; // defining 7 seconds timeout
+  // AniTubeApi api = AniTubeApi();
+  // var homePage;
+  // try {
+  //   // This call returns a HomePageInfo instance or throws an exception
+  //   // if somethings goes wrong,
+  //   homePage = await api.getHomePageData(timeout: localTimeout);
 
-    print('========== MOST RECENT ANIMES ==========');
-    homePage.mostRecentAnimes.forEach( (animeItem) => print );
-    print('========================================\n');
+  //   print('========== MOST RECENT ANIMES ==========');
+  //   homePage.mostRecentAnimes.forEach( (animeItem) => print );
+  //   print('========================================\n');
 
-    print('========== MOST SHOWED ANIMES ==========');
-    homePage.mostShowedAnimes.forEach( (animeItem) => print );
-    print('========================================\n');
+  //   print('========== MOST SHOWED ANIMES ==========');
+  //   homePage.mostShowedAnimes.forEach( (animeItem) => print );
+  //   print('========================================\n');
 
-    print('========== LATEST EPISODES =============');
-    homePage.latestEpisodes .forEach( (episodeItem) => print );
-    print('========================================\n');
+  //   print('========== LATEST EPISODES =============');
+  //   homePage.latestEpisodes .forEach( (episodeItem) => print );
+  //   print('========================================\n');
 
-  } on CrawlerApiException catch (ex){
-    print('${ex.errorType} -> ${ex.message}');
-  }
-
-
+  // } on CrawlerApiException catch (ex){
+  //   print('${ex.errorType} -> ${ex.message}');
+  // }
 }
