+++ lib/src/domain/irepository/ianime_details_repository.dart
@@ -1,0 +1,3 @@
+abstract class IAnimeDetailsRepository {
+  Future<String> getAnimeDetails(final String animeId);
+}
