+++ lib/src/domain/usecases/read_genres.dart
@@ -3,24 +3,21 @@ import 'package:anitube_crawler_api/src/domain/entities/parser/ihtml_parser.dart
 import 'package:anitube_crawler_api/src/domain/irepository/igenre_repository.dart';
 
 abstract class IReadGenres {
-  final IHTMLParser<List<Genre>> genrePageParser;
-
-  IReadGenres(this.genrePageParser);
+  IReadGenres();
   Future<List<Genre>> getGenres();
 }
 
 class ReadGenres extends IReadGenres {
   final IGenreRepository repository;
-
+  final IHTMLParser<List<Genre>> parser;
   ReadGenres(
     this.repository,
-    IHTMLParser<List<Genre>> parser,
-  )   : assert(repository != null),
-        super(parser);
+    this.parser,
+  ) : assert(repository != null);
 
   @override
   Future<List<Genre>> getGenres() async {
     final pageHTML = await repository.getGenres();
-    return this.genrePageParser.parseHTML(pageHTML);
+    return this.parser.parseHTML(pageHTML);
   }
 }
