+++ lib/src/external/anitube/parser/anitube_genres_parser.dart
@@ -1,30 +1,27 @@
-import 'package:anitube_crawler_api/anitube_crawler_api.dart';
 import 'package:anitube_crawler_api/src/domain/entities/genre.dart';
 import 'package:anitube_crawler_api/src/domain/entities/parser/ihtml_parser.dart';
+import 'package:anitube_crawler_api/src/domain/exceptions/CrawlerApiException.dart';
 import 'package:html/parser.dart';
 
 class AnitubeGenreParser implements IHTMLParser<List<Genre>> {
   @override
   List<Genre> parseHTML(String html) {
-    var list = <Genre>[];
+    if (html == null) throw ParserException(message: "null html data");
+    try {
+      var list = <Genre>[];
+      var body = parse(html);
 
-    if (html != null || html.isNotEmpty) {
-      try {
-        var body = parse(html);
-
-        body
-            .getElementsByClassName(_GenrePageNames.GENRE_CONTAINER)[0]
-            .children
-            .forEach((genreAnchor) {
-          list.add(Genre(genreAnchor.text));
-        });
-      } catch (ex) {
-        print('GenrePageParser error while parsing.\n $ex');
-        throw ParserException(message: "Error parsing genres data.");
-      }
+      body
+          .getElementsByClassName(_GenrePageNames.GENRE_CONTAINER)[0]
+          .children
+          .forEach((genreAnchor) {
+        list.add(Genre(genreAnchor.text));
+      });
+      return list;
+    } catch (ex) {
+      print('GenrePageParser error while parsing.\n $ex');
+      throw ParserException(message: "Error parsing genres data.");
     }
-
-    return list;
   }
 }
