+++ lib/src/infra/repository/genre_repository.dart
@@ -1,4 +1,3 @@
-import 'package:anitube_crawler_api/src/domain/entities/genre.dart';
 import 'package:anitube_crawler_api/src/domain/irepository/igenre_repository.dart';
 import 'package:anitube_crawler_api/src/infra/data_source/page_fetcher.dart';
