+++ test/mock_data/paths.dart
@@ -1,2 +1,5 @@
 const MOCKED_GENRES_HTML_PAGE =
     "./test/mock_data/anitube_pages/genres_page.html";
+
+const MOCKED_ANIME_DETAILS_HTML_PAGE =
+    "./test/mock_data/anitube_pages/anime_details_page.html";
